Added full set of methods for requesting an amount and the page number
Query methods start with the word query and include all of the same popular searches as the get methods. 

Query methods have two paramters that are optional. 
The first is an object specifying perPage and page where perPage is the amount of photos to get and page is the page to get the photos from.

The second object is the normal request options.  Since all methods ensure they do not send a request with null or undefined values and there is a default request options object, both query and get method parameters are optional here.